+++ src/admin/resources/js/cmsadmin.js
@@ -1137,6 +1137,7 @@
 			typeDataCopy.description = itemCopy.description;
 			typeDataCopy.keywords = itemCopy.keywords;
 			typeDataCopy.timestamp_create = itemCopy.timestamp_create;
+			typeDataCopy.image_id = itemCopy.image_id;
 			$http.post(
 				'admin/api-cms-navitem/update-page-item?navItemId=' + navItemId + '&navItemType=' + itemCopy.nav_item_type,
 				$.param(typeDataCopy),